+++ src/app/domain-annotations/prop.ts
@@ -24,7 +24,7 @@ namespace DomainAnnotations {
     }
 
     export function isApplicable(model?: Model): boolean {
-        return !!model && Model.isFromPdbArchive(model);
+        return !!model && Model.hasPdbId(model);
     }
 
     export function fromJson(model: Model, data: any) {