+++ src/app/ui/pdbe-structure-controls.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import { PluginUIComponent } from 'Molstar/mol-plugin-ui/base';
 import { StructureComponentControls } from 'Molstar/mol-plugin-ui/structure/components';
 import { StructureMeasurementsControls } from 'Molstar/mol-plugin-ui/structure/measurements';
@@ -7,7 +6,7 @@ import { VolumeStreamingControls, VolumeSourceControls } from 'Molstar/mol-plugi
 import { AnnotationsComponentControls } from './annotation-controls';
 import { Icon, BuildSvg } from 'Molstar/mol-plugin-ui/controls/icons';
 import { SuperpositionComponentControls } from './superposition-components';
-import { AFConfidenceComponentControls } from './af-confidence-controls';
+import { StructureQuickStylesControls } from 'Molstar/mol-plugin-ui/structure/quick-styles';
 
 export class PDBeStructureTools extends PluginUIComponent {
     render() {
@@ -16,11 +15,27 @@ export class PDBeStructureTools extends PluginUIComponent {
 
             <StructureSourceControls />
             <AnnotationsComponentControls />
+            <StructureQuickStylesControls />
             <StructureComponentControls />
             <VolumeStreamingControls />
             <VolumeSourceControls />
             <StructureMeasurementsControls />
+            <CustomStructureControls />
         </>;
+    }
+}
+
+export class CustomStructureControls extends PluginUIComponent<{ initiallyCollapsed?: boolean }> {
+    componentDidMount() {
+        this.subscribe(this.plugin.state.behaviors.events.changed, () => this.forceUpdate());
+    }
+
+    render() {
+        const controls: JSX.Element[] = [];
+        this.plugin.customStructureControls.forEach((Controls, key) => {
+            controls.push(<Controls initiallyCollapsed={this.props.initiallyCollapsed} key={key} />);
+        });
+        return controls.length > 0 ? <>{controls}</> : null;
     }
 }
 
@@ -31,6 +46,7 @@ export class PDBeLigandViewStructureTools extends PluginUIComponent {
             <StructureComponentControls />
             <VolumeStreamingControls />
             <StructureMeasurementsControls />
+            <CustomStructureControls />
         </>;
     }
 }
@@ -41,17 +57,7 @@ export class PDBeSuperpositionStructureTools extends PluginUIComponent {
             <div className='msp-section-header'><Icon svg={BuildSvg} />Structure Tools</div>
             <SuperpositionComponentControls />
             <StructureMeasurementsControls />
-        </>;
-    }
-}
-
-export class PDBeAfViewStructureTools extends PluginUIComponent {
-    render() {
-        return <>
-            <div className='msp-section-header'><Icon svg={BuildSvg} />Structure Tools</div>
-            <StructureComponentControls />
-            <AFConfidenceComponentControls />
-            <StructureMeasurementsControls />
+            <CustomStructureControls />
         </>;
     }
 }