+++ src/app/ui/segment-tree.tsx
@@ -703,8 +703,8 @@ class StructureRepresentationEntry extends PurePluginUIComponent<{ group: Struct
     render() {
         const repr = this.props.representation.cell;
         let label = repr.obj?.label;
-        if(repr.obj?.data.source && repr.obj?.data.source.label){
-            let sourceLabel = (repr.obj?.data.source.label.indexOf('[Focus]') >= 0) ? '[Focus]' : repr.obj?.data.source.label;
+        if(repr.obj?.data.repr && repr.obj?.data.repr.label){
+            let sourceLabel = (repr.obj?.data.repr.label.indexOf('[Focus]') >= 0) ? '[Focus]' : repr.obj?.data.repr.label;
             const isLargeLabel = sourceLabel.length > 10 ? true : false;
             sourceLabel = `${isLargeLabel ? `${sourceLabel.substring(0,28)}...` : sourceLabel}`;
             if(isLargeLabel) {
@@ -713,7 +713,7 @@ class StructureRepresentationEntry extends PurePluginUIComponent<{ group: Struct
                 label = `${sourceLabel} ${(label && label.length < 21) ? ' - ' + label : ''}`;
             }
         }
-        if(repr.obj?.data.source && repr.obj?.data.source.label === 'Custom Selection') label = 'Custom Selection';
+        if(repr.obj?.data.repr && repr.obj?.data.repr.label === 'Custom Selection') label = 'Custom Selection';
         return <div className='msp-representation-entry'>
             {repr.parent && <ExpandGroup header={`${label || 'Representation'}`} noOffset headerStyle={{ overflow: 'hidden' }} >
                 <UpdateTransformControl state={repr.parent} transform={repr.transform} customHeader='none' noMargin />